feat: add selected style for selected cards
