+++ src/components/game/game.js
@@ -2,60 +2,24 @@ import React, {Component} from "react";
 import Card from "../card/card";
 
 import './game.scss';
-import {Link} from "react-router-dom";
 
 export default class Game extends Component {
+
     render() {
-        const cards = [
-            {
-                name: 'shell',
-                img: 'images/blueshell.png',
-            },
-            {
-                name: 'star',
-                img: 'images/star.png',
-            },
-            {
-                name: 'bobomb',
-                img: 'images/bobomb.png',
-            },
-            {
-                name: 'mario',
-                img: 'images/mario.png',
-            },
-            {
-                name: 'luigi',
-                img: 'images/luigi.png',
-            },
-            {
-                name: 'peach',
-                img: 'images/peach.png',
-            },
-            {
-                name: '1up',
-                img: 'images/1up.png',
-            },
-            {
-                name: 'mushroom',
-                img: 'images/mushroom.png',
-            },
-            {
-                name: 'thwomp',
-                img: 'images/thwomp.png',
-            },
-            {
-                name: 'bulletbill',
-                img: 'images/bulletbill.png',
-            },
-            {
-                name: 'coin',
-                img: 'images/coin.png',
-            },
-            {
-                name: 'goomba',
-                img: 'images/goomba.png',
-            },
-        ];
+        const {cards} = this.props;
+
+        const twoPacksOfCards = [...cards, ...cards];
+        let i = 0;
+
+        const allCards = twoPacksOfCards.map(({name, img}) => {
+            return (
+                <Card className={`card`}
+                      key={i++}
+                      name={name}
+                      style={{backgroundImage: `url(${img})`}}/>
+            );
+        });
+
         return (
             <section className='game'>
                 <section className="status">
@@ -69,7 +33,7 @@ export default class Game extends Component {
                     </div>
                 </section>
                 <section className="game-box">
-                    <Card cards={cards}/>
+                    {allCards}
                 </section>
                 <div className="buttons">
                     <button className="button button__new-game" type="button">New Game</button>
